Reduce serialization overhead by ~40% by using 8/16/32 bit ints as needed
Also, fix a bug in deserializing the animation repeat sequence.

+++ include/game/protos/object.h
@@ -52,12 +52,12 @@ struct object_t {
     vec2f vel;
     int vstate;
     int hstate;
-    int direction;
+    int8_t direction;
     float y_percent;
     float gravity;
     
-    int group;
-    int layers;
+    int8_t group;
+    uint8_t layers;
     
     int cur_animation_own;
     animation *cur_animation;
@@ -68,7 +68,7 @@ struct object_t {
     palette *cur_palette;
     int cur_remap;
     int halt;
-    int stride;
+    uint8_t stride;
     int cast_shadow;
     texture *cur_texture;
